functional parser
This file still needs a lot of cleaning from my last attempts, but it fully reads and parses the svg files now. I also have some methods of just displaying the SVGs commented out at the top. Next is to make sure it works with other files and adjust it so it can be iterated more cleanly